Add files via upload
Task 1
I created a new report package and created three classes. I created a ReportingStructure Object to hold the employee object as well as the calculated number of direct reports. Inside the ReportingStructure Object I calculate the number of direct reports by getting the size of the directReport list and storing that in the numberOfReports variable. In the Rest Controller I created the getDirectReports function which takes in the employee Id and makes a call to the ReportService class where I make a call to the EmployeeRepository to retrieve the associated employee. Once I have retrieved the employee I return a newly created ReportingStructure object which outputs the employee data as well as the direct reports. Below is an example of the output:

{
    "employee": {
        "employeeId": "03aa1462-ffa9-4978-901b-7c001562cf6f",
        "firstName": "Ringo",
        "lastName": "Starr",
        "position": "Developer V",
        "department": "Engineering",
        "directReports": [
            {
                "employeeId": "62c1084e-6e34-4630-93fd-9153afb65309",
                "firstName": null,
                "lastName": null,
                "position": null,
                "department": null,
                "directReports": null
            },
            {
                "employeeId": "c0c2293d-16bd-4603-8e08-638a9d18b22c",
                "firstName": null,
                "lastName": null,
                "position": null,
                "department": null,
                "directReports": null
            },
            {
                "employeeId": "c0c2293d-16bd-4603-8e08-638a9d18b22c",
                "firstName": null,
                "lastName": null,
                "position": null,
                "department": null,
                "directReports": null
            },
            {
                "employeeId": "c0c2293d-16bd-4603-8e08-638a9d18b22c",
                "firstName": null,
                "lastName": null,
                "position": null,
                "department": null,
                "directReports": null
            }
        ]
    },
    "numberOfReports": 4
}